sensors/dpi_event_forwarder/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.1
watchdog==3.0.0